requirements.txt
locust>=2.30.0
pyyaml>=6.0.1
beautifulsoup4>=4.12.3
